better way to set autocast maybe

+++ scripts/v0.2_sweep/latent_mim_no_worldcover.py
@@ -2,6 +2,7 @@
 
 import logging
 
+from olmo_core.config import DType
 from olmo_core.distributed.parallel.data_parallel import (
     DataParallelConfig,
     DataParallelType,
@@ -103,7 +104,11 @@ def build_train_module_config(
         max_grad_norm=1.0,
         scheduler=CosWithWarmup(),
         ema_decay=(1.0, 1.0),
-        dp_config=DataParallelConfig(name=DataParallelType.fsdp),
+        dp_config=DataParallelConfig(
+            name=DataParallelType.fsdp,
+            param_dtype=DType.bfloat16,
+            reduce_dtype=DType.float32,
+        ),
     )
